refactor: consolidate benchmarks into a single-file poc/benches crate #105
