+++ poc/README.md
@@ -7,18 +7,22 @@ This directory contains proof-of-concept implementations of the Schema-Segment C
 ```text
 poc/
 ├── standard/               Main Rust workspace (host CPU simulation)
-│   ├── Cargo.toml          Workspace manifest with 18 member crates
+│   ├── Cargo.toml          Workspace manifest with 17 member crates
 │   ├── crates/
 │   │   ├── core/           Absolute primitives (Segment, Field, Projector trait)
 │   │   ├── primitive/      Scheme abstraction layer
 │   │   ├── schemes/        Concrete Scheme implementations
 │   │   ├── examples/       Shared utilities (projectors, parser, compiler pipeline, emitters)
-│   │   ├── benchmarks/     Validation-only kernels: Rust baseline vs SSCCS formulation
 │   │   ├── concepts/    11 independent constitutional concept test crates
 │   │   ├── field-synthesis/   Research placeholder
 │   │   ├── hardware-mapping/  Research placeholder
 │   │   └── compiler-opt/      Research placeholder
 │   └── README.md           Full crate-level documentation
+│
+├── benches/                Standalone validation-only benchmark crate
+│   ├── Cargo.toml          ssccs-benchmarks (criterion)
+│   ├── bench.rs            Single-file kernels: Rust baseline vs SSCCS formulation
+│   └── run.sh              Benchmark runner: verify + kernels, results in result/
 │
 ├── baremetal_riscv/        RISC-V assembly integration crate
 │   ├── Cargo.toml          Host build; RV64 assembly path under Spike
@@ -31,7 +35,7 @@ poc/
 │   ├── index.md            Human-readable documentation index
 │   └── *.md                Per-crate API reference (generated by doc.py)
 │
-├── run.sh                  Full validation: fmt → clippy → build → test → run
+├── run.sh                  Full validation: fmt → clippy → build → test → run → bench
 ├── doc.py                  Generates crate documentation into _llms/
 └── README.md               This file
 ```
@@ -48,7 +52,6 @@ Rust workspace targeting x86_64 / aarch64. Implements the complete SSCCS ontolog
 | Scheme | `ssccs-primitive` | Scheme, SchemeBuilder, SchemeTrait, Axis, StructuralRelation, MemoryLayout, ObservationRules |
 | Scheme | `ssccs-schemes` | Grid2DTemplate, IntegerLineTemplate, GraphTemplate, Tensor3DTemplate, CompositeScheme, TransformedScheme, BooleanSpace, IntegerSpace |
 | Utilities | `ssccs-examples` | IntegerProjector, ArithmeticProjector, ParityProjector, CoordinateSumProjector, CompilerPipeline, .ss parser, assembly data emitter, branchless constraint gate emitter |
-| Benchmarks | `ssccs-benchmarks` | Vector addition, 2D convolution, graph BFS: Rust baseline vs SSCCS formulation (validation-only) |
 | Experiments | `concept-segment` through `concept-integrated` plus `experiment-data-processing` | Each crate tests one constitutional concept independently |
 | Research | `ssccs-field-synthesis` | Field composition algebra (placeholder) |
 | Research | `ssccs-hardware-mapping` | Scheme-to-hardware mapping (placeholder) |
@@ -152,7 +155,7 @@ Generated gates follow `fn(*const i64) -> u32`: coordinate pointer in `a0`, gate
 
 - Determinism and race-free concurrent observation tests in `ssccs-examples/tests/`.
 - Assembly syntax gate over all `asm/*.S`, Spike execution, and SystemVerilog cross-validation in `baremetal_riscv/`.
-- Benchmark kernels in `ssccs-benchmarks` (run `cargo bench -p ssccs-benchmarks`): the measured emulation overhead on existing hardware informs the practical plane; it never shapes the model.
+- Benchmark kernels in `benches/` (run `benches/run.sh`): the measured emulation overhead on existing hardware informs the practical plane; it never shapes the model.
 
 ## License
