+++ poc/benches/Cargo.toml
@@ -5,12 +5,13 @@ edition = "2024"
 description = "Validation-only benchmarking kernels: pure Rust baseline vs SSCCS reference simulation"
 
 [dependencies]
-ssccs-core = { path = "../core" }
-ssccs-primitive = { path = "../primitive" }
+ssccs-core = { path = "../standard/crates/core" }
+ssccs-primitive = { path = "../standard/crates/primitive" }
 
 [dev-dependencies]
 criterion = "0.7"
 
 [[bench]]
-name = "kernels"
+name = "bench"
+path = "bench.rs"
 harness = false
