+++ poc/standard/Cargo.toml
@@ -5,7 +5,6 @@ members = [
     "crates/primitive",
     "crates/schemes",
     "crates/examples",
-    "crates/benchmarks",
     "crates/field-synthesis",
     "crates/hardware-mapping",
     # Concept crates — each implements and validates one SSCCS concept
